information-retrieval-api/pyproject.toml
[project]
name = "information-retrieval-api"
version = "0.1.0"
description = "An API for information retrieval tasks."
requires-python = ">=3.12"
dependencies = [
    "aiocache>=0.12.3",
    "bs4>=0.0.2",
    "fastapi>=0.135.2",
    "greenlet>=3.1.1",
    "gunicorn>=25.2.0",
    "marshmallow>=4.2.2",
    "matplotlib>=3.10.8",
    "nltk>=3.9.4",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
    "asyncpg>=0.30.0",
    "scikit-learn>=1.8.0",
    "selenium>=4.41.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
